+++ examples/02.stack-blow-up.php
@@ -6,35 +6,45 @@ use VovanVE\HtmlTemplate\caching\memory\CacheStrings;
 use VovanVE\HtmlTemplate\compile\Compiler;
 use VovanVE\HtmlTemplate\Engine;
 use VovanVE\HtmlTemplate\runtime\RuntimeHelper;
+use VovanVE\HtmlTemplate\runtime\RuntimeHelperInterface;
 use VovanVE\HtmlTemplate\source\memory\TemplateStringProvider;
+
+class FooComponent extends \VovanVE\HtmlTemplate\components\BaseComponent
+{
+    /** @var string */
+    public $level;
+
+    /**
+     * @param RuntimeHelperInterface $runtime
+     * @param \Closure|null $content
+     * @return string
+     */
+    public function render(RuntimeHelperInterface $runtime, ?\Closure $content = null): string
+    {
+        if (null === $content) {
+            return "";
+        }
+        return join('', $content($runtime));
+    }
+}
+
+$input = 'From the depth.';
+for ($i = 500; $i > 0; $i--) {
+    $input = "<Foo level='$i'>$input</Foo>";
+}
 
 $engine = (new Engine)
     ->setCache(new CacheStrings('CacheTempRuntime_%{hash}'))
     ->setTemplateProvider(
         (new TemplateStringProvider)
-            ->setTemplate(
-                'foo',
-                <<<'TPL'
-<a href={ link } title={ "Foo bar: ${ title }" }>
-    <span id=foobar class='it parses html'>
-        { description }
-    </span>
-    <span>
-        {# comment #}
-    </span>
-</a>
-TPL
-            )
+            ->setTemplate('foo', $input)
     )
     ->setCompiler(new Compiler());
 
-$runtime = new RuntimeHelper([
-    'link' => 'http://example.com?foo=bar&lorem=ipsum#hash',
-    'title' => 'Lorem <ipsum> "dolor" sit amet',
-    'description' => function () {
-        return 'Some <text/plain> content';
-    },
-]);
+$runtime = (new RuntimeHelper)
+    ->addComponents([
+        'Foo' => FooComponent::class,
+    ]);
 
 $compiled = $engine->compileTemplate('foo');