+++ src/compile/Compiler.php
@@ -20,7 +20,6 @@ use VovanVE\HtmlTemplate\report\Message;
 use VovanVE\HtmlTemplate\report\MessageInterface;
 use VovanVE\HtmlTemplate\report\Report;
 use VovanVE\HtmlTemplate\report\ReportInterface;
-use VovanVE\HtmlTemplate\runtime\RuntimeHelper;
 use VovanVE\HtmlTemplate\source\TemplateInterface;
 use VovanVE\parser\actions\ActionAbortException;
 use VovanVE\parser\actions\ActionsMadeMap;
@@ -32,7 +31,7 @@ class Compiler implements CompilerInterface
 {
     private const A_BUBBLE = Parser::ACTION_BUBBLE_THE_ONLY;
 
-    private const VERSION = '0.1.2';
+    private const VERSION = '0.2.0';
 
     private const STRING_ESCAPE_LETTER = [
         'b' => "\x08",
@@ -311,7 +310,7 @@ class Compiler implements CompilerInterface
             return $new_list;
         };
         $htmlDecodeNow = function (string $content) {
-            return RuntimeHelper::htmlDecodeEntity($content);
+            return CompilerHelper::htmlDecodeEntity($content);
         };
         $concatTwoFragments = function (string $a, string $b) {
             return $a . $b;