+++ src/runtime/RuntimeTemplateException.php
@@ -1,0 +1,6 @@
+<?php
+namespace VovanVE\HtmlTemplate\runtime;
+
+class RuntimeTemplateException extends \Exception
+{
+}