+++ tests/helpers/CounterStepComponent.php
@@ -2,21 +2,23 @@
 namespace VovanVE\HtmlTemplate\tests\helpers;
 
 use VovanVE\HtmlTemplate\components\BaseComponent;
+use VovanVE\HtmlTemplate\runtime\RuntimeHelperInterface;
 
 class CounterStepComponent extends BaseComponent
 {
     public $step;
 
     /**
-     * @param array|null $content
+     * @param RuntimeHelperInterface $runtime
+     * @param \Closure|null $content
      * @return string
      */
-    public function render(?array $content = null): string
+    public function render(RuntimeHelperInterface $runtime, ?\Closure $content = null): string
     {
         $marker = "step: " . ($this->step ?? '?');
         if (null === $content) {
             return "<!-- $marker /-->";
         }
-        return "<!-- $marker -->" . join('', $content) . "<!-- /$marker -->";
+        return "<!-- $marker -->" . join('', $content($runtime)) . "<!-- /$marker -->";
     }
 }