+++ tests/helpers/TestComponent.php
@@ -2,6 +2,7 @@
 namespace VovanVE\HtmlTemplate\tests\helpers;
 
 use VovanVE\HtmlTemplate\components\BaseComponent;
+use VovanVE\HtmlTemplate\runtime\RuntimeHelperInterface;
 
 class TestComponent extends BaseComponent
 {
@@ -9,10 +10,11 @@ class TestComponent extends BaseComponent
     public $bar;
 
     /**
-     * @param array|null $content
+     * @param RuntimeHelperInterface $runtime
+     * @param \Closure|null $content
      * @return string
      */
-    public function render(?array $content = null): string
+    public function render(RuntimeHelperInterface $runtime, ?\Closure $content = null): string
     {
         $head = "Test Component";
         $props = "foo=" . json_encode($this->foo) . " bar=" . json_encode($this->bar);
@@ -20,6 +22,6 @@ class TestComponent extends BaseComponent
         if (null === $content) {
             return "<!-- $head: $props /-->";
         }
-        return "<!-- $head: $props -->" . join('', $content) . "<!-- /$head -->";
+        return "<!-- $head: $props -->" . join('', $content($runtime)) . "<!-- /$head -->";
     }
 }